+++ stories/MigrationGuides/UpgradeTov3.mdx
@@ -1,0 +1,6 @@
+import Docs from "../../docs/upgrade-to-3.0.md?raw"
+import {Markdown, Meta} from "@storybook/blocks"
+
+<Meta title="Migration Guides/Upgrade to 3.0" />
+
+<Markdown>{Docs}</Markdown>
